+++ Tests/XcodeGenKitTests/SourceGeneratorTests.swift
@@ -602,7 +602,7 @@ class SourceGeneratorTests: XCTestCase {
 
                 let target = Target(name: "Test", type: .framework, platform: .iOS, sources: [
                     TargetSource(path: "A", buildPhase: .resources),
-                    TargetSource(path: "B", buildPhase: TargetSource.BuildPhase.none),
+                    TargetSource(path: "B", buildPhase: BuildPhaseSpec.none),
                     TargetSource(path: "C", buildPhase: nil),
                 ])
                 let project = Project(basePath: directoryPath, name: "Test", targets: [target])
@@ -614,11 +614,11 @@ class SourceGeneratorTests: XCTestCase {
                 try pbxProj.expectFile(paths: ["A", "Info.plist"], buildPhase: .resources)
                 try pbxProj.expectFile(paths: ["A", "file.xcconfig"], buildPhase: .resources)
 
-                try pbxProj.expectFile(paths: ["B", "file.swift"], buildPhase: TargetSource.BuildPhase.none)
-                try pbxProj.expectFile(paths: ["B", "file.xcassets"], buildPhase: TargetSource.BuildPhase.none)
-                try pbxProj.expectFile(paths: ["B", "file.h"], buildPhase: TargetSource.BuildPhase.none)
-                try pbxProj.expectFile(paths: ["B", "Info.plist"], buildPhase: TargetSource.BuildPhase.none)
-                try pbxProj.expectFile(paths: ["B", "file.xcconfig"], buildPhase: TargetSource.BuildPhase.none)
+                try pbxProj.expectFile(paths: ["B", "file.swift"], buildPhase: BuildPhaseSpec.none)
+                try pbxProj.expectFile(paths: ["B", "file.xcassets"], buildPhase: BuildPhaseSpec.none)
+                try pbxProj.expectFile(paths: ["B", "file.h"], buildPhase: BuildPhaseSpec.none)
+                try pbxProj.expectFile(paths: ["B", "Info.plist"], buildPhase: BuildPhaseSpec.none)
+                try pbxProj.expectFile(paths: ["B", "file.xcconfig"], buildPhase: BuildPhaseSpec.none)
 
                 try pbxProj.expectFile(paths: ["C", "file.swift"], buildPhase: .sources)
                 try pbxProj.expectFile(paths: ["C", "file.m"], buildPhase: .sources)
@@ -633,16 +633,16 @@ class SourceGeneratorTests: XCTestCase {
                 try pbxProj.expectFile(paths: ["C", "file.tpp"], buildPhase: .headers)
                 try pbxProj.expectFile(paths: ["C", "file.hxx"], buildPhase: .headers)
                 try pbxProj.expectFile(paths: ["C", "file.def"], buildPhase: .headers)
-                try pbxProj.expectFile(paths: ["C", "file.xcconfig"], buildPhase: TargetSource.BuildPhase.none)
-                try pbxProj.expectFile(paths: ["C", "file.entitlements"], buildPhase: TargetSource.BuildPhase.none)
-                try pbxProj.expectFile(paths: ["C", "file.gpx"], buildPhase: TargetSource.BuildPhase.none)
-                try pbxProj.expectFile(paths: ["C", "file.apns"], buildPhase: TargetSource.BuildPhase.none)
-                try pbxProj.expectFile(paths: ["C", "file.xcconfig"], buildPhase: TargetSource.BuildPhase.none)
-                try pbxProj.expectFile(paths: ["C", "file.xcconfig"], buildPhase: TargetSource.BuildPhase.none)
-                try pbxProj.expectFile(paths: ["C", "file.xcconfig"], buildPhase: TargetSource.BuildPhase.none)
+                try pbxProj.expectFile(paths: ["C", "file.xcconfig"], buildPhase: BuildPhaseSpec.none)
+                try pbxProj.expectFile(paths: ["C", "file.entitlements"], buildPhase: BuildPhaseSpec.none)
+                try pbxProj.expectFile(paths: ["C", "file.gpx"], buildPhase: BuildPhaseSpec.none)
+                try pbxProj.expectFile(paths: ["C", "file.apns"], buildPhase: BuildPhaseSpec.none)
+                try pbxProj.expectFile(paths: ["C", "file.xcconfig"], buildPhase: BuildPhaseSpec.none)
+                try pbxProj.expectFile(paths: ["C", "file.xcconfig"], buildPhase: BuildPhaseSpec.none)
+                try pbxProj.expectFile(paths: ["C", "file.xcconfig"], buildPhase: BuildPhaseSpec.none)
                 try pbxProj.expectFile(paths: ["C", "file.xcassets"], buildPhase: .resources)
                 try pbxProj.expectFile(paths: ["C", "file.123"], buildPhase: .resources)
-                try pbxProj.expectFile(paths: ["C", "Info.plist"], buildPhase: TargetSource.BuildPhase.none)
+                try pbxProj.expectFile(paths: ["C", "Info.plist"], buildPhase: BuildPhaseSpec.none)
                 try pbxProj.expectFile(paths: ["C", "file.metal"], buildPhase: .sources)
                 try pbxProj.expectFile(paths: ["C", "file.mlmodel"], buildPhase: .sources)
                 try pbxProj.expectFile(paths: ["C", "Intent.intentdefinition"], buildPhase: .sources)
@@ -652,6 +652,57 @@ class SourceGeneratorTests: XCTestCase {
                 try pbxProj.expectFileMissing(paths: ["C", "Settings.bundle", "Root.plist"])
                 try pbxProj.expectFileMissing(paths: ["C", "WithPeriod2.0"])
                 try pbxProj.expectFile(paths: ["C", "WithPeriod2.0", "file.swift"], buildPhase: .sources)
+            }
+
+            $0.it("sets file type properties") {
+                let directories = """
+                  A:
+                    - file.resource1
+                    - file.source1
+                    - file.abc:
+                        - file.a
+                    - file.exclude1
+                    - file.unphased1
+                    - ignored.swift
+                """
+                try createDirectories(directories)
+
+                let target = Target(name: "Test", type: .framework, platform: .iOS, sources: [
+                    TargetSource(path: "A"),
+                ])
+                let project = Project(basePath: directoryPath, name: "Test", targets: [target], options: .init(fileTypes: [
+                    "abc": FileType(buildPhase: .sources),
+                    "source1": FileType(buildPhase: .sources, attributes: ["a1", "a2"], resourceTags: ["r1", "r2"], compilerFlags: ["-c1", "-c2"]),
+                    "resource1": FileType(buildPhase: .resources, attributes: ["a1", "a2"], resourceTags: ["r1", "r2"], compilerFlags: ["-c1", "-c2"]),
+                    "unphased1": FileType(buildPhase: BuildPhaseSpec.none),
+                    "swift": FileType(buildPhase: .resources),
+                ]))
+
+                let pbxProj = try project.generatePbxProj()
+                try pbxProj.expectFile(paths: ["A", "file.abc"], buildPhase: .sources)
+                try pbxProj.expectFile(paths: ["A", "file.source1"], buildPhase: .sources)
+                try pbxProj.expectFile(paths: ["A", "file.resource1"], buildPhase: .resources)
+                try pbxProj.expectFile(paths: ["A", "file.unphased1"], buildPhase: BuildPhaseSpec.none)
+                try pbxProj.expectFile(paths: ["A", "ignored.swift"], buildPhase: .resources)
+
+                do {
+                    let fileReference = try unwrap(pbxProj.getFileReference(paths: ["A", "file.resource1"], names: ["A", "file.resource1"]))
+                    let buildFile = try unwrap(pbxProj.buildFiles.first(where: { $0.file === fileReference }))
+                    let settings = NSDictionary(dictionary: buildFile.settings ?? [:])
+                    try expect(settings) == [
+                        "ATTRIBUTES": ["a1", "a2"],
+                        "ASSET_TAGS": ["r1", "r2"],
+                    ]
+                }
+                do {
+                    let fileReference = try unwrap(pbxProj.getFileReference(paths: ["A", "file.source1"], names: ["A", "file.source1"]))
+                    let buildFile = try unwrap(pbxProj.buildFiles.first(where: { $0.file === fileReference }))
+                    let settings = NSDictionary(dictionary: buildFile.settings ?? [:])
+                    try expect(settings) == [
+                        "ATTRIBUTES": ["a1", "a2"],
+                        "COMPILER_FLAGS": "-c1 -c2",
+                        ]
+                }
             }
 
             $0.it("duplicate TargetSource is included once in sources build phase") {
@@ -1084,7 +1135,7 @@ class SourceGeneratorTests: XCTestCase {
 extension PBXProj {
 
     /// expect a file within groups of the paths, using optional different names
-    func expectFile(paths: [String], names: [String]? = nil, buildPhase: TargetSource.BuildPhase? = nil, file: String = #file, line: Int = #line) throws {
+    func expectFile(paths: [String], names: [String]? = nil, buildPhase: BuildPhaseSpec? = nil, file: String = #file, line: Int = #line) throws {
         guard let fileReference = getFileReference(paths: paths, names: names ?? paths) else {
             var error = "Could not find file at path \(paths.joined(separator: "/").quoted)"
             if let names = names, names != paths {